requirements.txt
groq>=0.4.0
aiokafka>=0.8.0
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
prometheus-client>=0.16.0
python-dotenv>=1.0.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
